Update to suggest that brew is used
My experience has been that downloading the chrome webdriver manually on a Mac is fraught with issues, especially when chrome is updated. 

There is now the ability to install the chrome webdriver via Homebrew, so I think we should suggest this in the docs, as it works out of the box for me.